fix for open shapes when using tool compensation
adding G40 before each G41/G42 invokation is a dirty fix for open shapes
when using tool compensation. we need to turn it off before turning it
on in the opposite direction so the linuxCNC axis does not complain.
Needs more elegant way to do that in the code. (see closed ticket 57)