Fixed but during export with Cutter_Compensation done by machine
shape.py:
If there is no offshape creted nothing will be exportet for this shape.
However for the Route optimisation this is still considered. 
FIXME: Route Optimisation not optimal on this point

stmove.py:
Needed to change the call of the shapeoffset.py functin inside of the
stmove.py in order to have it expected way for the function.