Updates according to patches created by Richard:
I have created a patch for the 20170925 version to include both upper
and lower case filenames in the open dialogue (attached).  The patch
also includes updated translation files to match the change.

1. the tool number is read from the layer description as a float.  As
the tool number is converted to string before comparison with the tool
table it end up with a trailing decimal point (e.g. 1.0 rather than 1)
for the comparison - which can therefore never match.
2. If the tool does match the rest of the parameters - diameter, feed
etc. - are not updated (unless the tool table itself is subsequently
changed) and cannot be changed from the user interface since the tool
number is defined.