optionally skip running pylupdate to update the QT Linguist translation files
When packaging dxf2gcode for Debian it makes no sense to run pylupdate
to update the database of translatable strings, since no new translations
are about to be added.

Furthermore, changing the *.ts files at build time introduces problems
with build repeatability.

This commit introduces an optional --no-pylupdate command-line argument
to make_tr.py which skips the problematic pylupdate step.  The Debian
build scripts pass in this argument when needed.  If this argument is
*not* passed in, make_tr.py behaves exactly like before, ie it calls
pylupdate to update the translation files.