Processing for a PLASMA named layer
Added concept of PLASMA named layer to layercontent.  This is in prep for any specific overrides that make sense to be able to be processed.

Added PLASMA layer specific processing to how autotoolcompensation is processed.  Plasma cutting for best results cares about the direction.   External cuts should be CW and internal cut-outs should be CCW.  This allows for the natural CW swirl of the plasma plume.  This "climb milling" approach against the desired cut line gives a better finish.