Changed a couple of standards/defaults based on advice received
[1] Changed default leadin and leadout to be Line
[2] Changed Line angle of approach to be perpendicular (90 deg) to the entry point.  Based on advice from a more experienced cnc plasma user RodW from Linuxcnc forums.  Same advice can be found on line from Hypertherm specialist Jim Colt (now retired).

ToDo: Just acknowledging that still need to provide a way to adjust/override the approach angle for those who want to.  Easiest way will be via dxf layer annotations. That way you can do different things based on the geometry in the layer.  Also the leadin/out length should be adjustable via layer annotations.  Down side is this is not portable between material thicknesses. Doing this in the UI will be a lot more work. Maybe in the future ....