Ability to adjust Lead-In/Out angle by shape via menu or keyboard
This is an initial implementation of an ability to refine the lead-in and lead-out angle of a path in the CAM side cutter compensation.  This allows more refined leads setting for shapes.

There is some odd behaviour in how the lead out behaves.  Also need to check if these settings are saved as part of the d2p file format.