Adding initial UI setting control of join colinear lines
Join Co-Linear Lines function needs to be mutually exclusive of split-line-segments.  It also fights with custom splits in geom in the dxf file intended ti support specific placement of lead ins.

This is an initial commit.  I am not sure it is fully working and I also need to added in mutually exclusive tests with split line segs as the two functions just fight each other to no good outcome.